pyproject.toml
[project]
name = "insurance-recommendation-agent"
version = "0.1.0"
description = "保险智能推荐Agent系统 - 基于LangGraph多Agent架构的智能对话推荐系统"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.2.0",
    "faiss-cpu>=1.9.0",
    "hypothesis>=6.119.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "alembic>=1.14.0",
    "sqlalchemy>=2.0.36",
    "numpy>=2.2.0",
    "sentence-transformers>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "locust>=2.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "skills", "tools", "api", "models", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
